test: add default editor configuration and tests for editor panel functionality
